// graphql/generated/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
};
// ...
export type Query = {
  __typename?: 'Query';
  tasks: Array<Task>;
  user?: Maybe<User>;
};


export type QueryUserArgs = {
  id: Scalars['Int']['input'];
};

export type Subtask = {
  __typename?: 'Subtask';
  completed: Scalars['Boolean']['output'];
  deadline: Scalars['DateTime']['output'];
  description: Scalars['String']['output'];
  id: Scalars['Int']['output'];
  name: Scalars['String']['output'];
  responsible: User;
};

export type Task = {
  __typename?: 'Task';
  id: Scalars['Int']['output'];
  subtasks: Array<Subtask>;
  title: Scalars['String']['output'];
};

export type User = {
  __typename?: 'User';
  email: Scalars['String']['output'];
  id: Scalars['Int']['output'];
  tasks?: Maybe<Array<Task>>;
  username: Scalars['String']['output'];
};

export type GetTasksQueryVariables = Exact<{ [key: string]: never; }>;


export type GetTasksQuery = { __typename?: 'Query', tasks: Array<{ __typename?: 'Task', id: number, title: string, subtasks: Array<{ __typename?: 'Subtask', id: number, name: string, deadline: any, completed: boolean, description: string, responsible: { __typename?: 'User', id: number, username: string, email: string } }> }> };